Assignment1 Go
Instructions

In this challenge, you will practice C# scripting with the MonoBehaviour class.

In a new file, create a new scene called "TestScene". In this scene, create a new cube and attach a new component called "TestBehaviour" to it. Make the cube move back and forth between multiple points, moving at a different velocity on each segment.

The position of the GameObject and its velocity on each segment must be configurable in the inspector.
Deliverables

    Script (a .cs file): MonoBehaviour class that moves the cube
    Short video (10 to 20 seconds): show how the cube moves.